Add nonbonded exclusions in NumberOfContacts CV
